ReadableStream read all Stash Data
Fix for the problem where the page does not reload and fill the form: The data that comes back from your Stash instance is returned as a ReadableStream and when the read() method is called on it the data is not necessarily complete, it's just whatever is in the buffer of the ReadableStream. This works fine for the smaller messages that basically just have the status of "Uploading Images" etc, but for the actual payload of the data the buffer often does not contain the entire set of data returned from Stash.